getting started documentation and prep for
