+++ docs/index.html
@@ -651,6 +651,7 @@
               {value: "directive-repeat.html", label: "Repeat"},
               {value: "directive-repeat-as.html", label: "Repeat As"},
               {value: "directive-attr.html", label: "Attr"},
+              {value: "directive-bind.html", label: "Bind"},
             ]
           },
           {
