+++ src/directives/bind.directive.ts
@@ -1,13 +1,14 @@
 import {Directive} from "../directive";
+import {CWCO} from "../cwco";
 
 export class Bind extends Directive {
 	parseValue(value: string, prop: string | null) {
-		return `['${(prop || '').trim()}', "${value}"]`;
+		return `["${(prop || '').trim()}", "${value}"]`;
 	}
 
-	render([prop, value]: [string, any], {element, anchorNode}: directiveRenderOptions) {
+	render([prop, value]: [string, any], {element, anchorNode}: CWCO.directiveRenderOptions) {
 		if (prop) {
-			(element as ObjectLiteral)[prop] = value;
+			(element as CWCO.ObjectLiteral)[prop] = value;
 		} else {
 			element.textContent = value;
 		}
