+++ src/directives/registry.ts
@@ -1,1 +1,3 @@
-export const directiveRegistry: {[key: string]: DirectiveConstructor} = {};
+import {CWCO} from "../cwco";
+
+export const directiveRegistry: {[key: string]: CWCO.DirectiveConstructor} = {};
