+++ src/utils/define-node-context-metadata.ts
@@ -1,22 +1,23 @@
 import {$} from "../metadata";
+import {CWCO} from "../cwco";
 
 export function defineNodeContextMetadata(node: Node) {
 	if ($.has(node) && $.get(node)?.$context) {
 		return;
 	}
 
-	let ctx: ObjectLiteral = {};
-	let subs: Array<ObserverCallback> = [];
-	const dt: ObjectLiteral = $.get(node) || {};
+	let ctx: CWCO.ObjectLiteral = {};
+	let subs: Array<CWCO.ObserverCallback> = [];
+	const dt: CWCO.ObjectLiteral = $.get(node) || {};
 
-	dt.subscribe = (cb: ObserverCallback) => {
+	dt.subscribe = (cb: CWCO.ObserverCallback) => {
 		subs.push(cb);
 		return () => {
 			subs = subs.filter((c) => c !== cb);
 		}
 	}
 
-	dt.updateContext = (newCtx: ObjectLiteral | null = null) => {
+	dt.updateContext = (newCtx: CWCO.ObjectLiteral | null = null) => {
 		if (newCtx && typeof newCtx === 'object') {
 			ctx = {...ctx, ...newCtx};
 		}
@@ -38,7 +39,7 @@ export function defineNodeContextMetadata(node: Node) {
 	})
 
 	function notify() {
-		((node as WebComponent).root ?? node).childNodes
+		((node as CWCO.WebComponent).root ?? node).childNodes
 			.forEach((n) => {
 				if (typeof $.get(n)?.updateContext === 'function') {
 					$.get(n).updateContext();
