+++ src/utils/get-event-handler-function.ts
@@ -1,6 +1,7 @@
 import {getComponentNodeEventListener} from "./get-component-node-event-listener";
+import {CWCO} from "../cwco";
 
-export function getEventHandlerFunction(component: WebComponent, nodeData: ObjectLiteral, attribute: Attr): EventListenerCallback | null {
+export function getEventHandlerFunction(component: CWCO.WebComponent, nodeData: CWCO.ObjectLiteral, attribute: Attr): CWCO.EventListenerCallback | null {
 	const props = Array.from(new Set([...Object.getOwnPropertyNames(nodeData), ...component.$properties]));
 	const values = props.map(k => {
 		return nodeData[k] ?? component[k] ?? null;
