+++ src/utils/parse-node-directive.ts
@@ -1,4 +1,6 @@
-export function parseNodeDirective(node: Element, name: string, value: string): DirectiveValue{
+import {CWCO} from "../cwco";
+
+export function parseNodeDirective(node: Element, name: string, value: string): CWCO.DirectiveValue {
 	const dot = name.indexOf('.');
 	let prop = null;
