+++ src/utils/resolve-executable.ts
@@ -1,6 +1,7 @@
 import {evaluateStringInComponentContext} from "./evaluate-string-in-component-context";
+import {CWCO} from "../cwco";
 
-export function resolveExecutable(component: WebComponent, nodeData: ObjectLiteral, {match, executable}: Executable, newValue: string) {
+export function resolveExecutable(component: CWCO.WebComponent, nodeData: CWCO.ObjectLiteral, {match, executable}: CWCO.Executable, newValue: string) {
 	let res = evaluateStringInComponentContext(executable, component, nodeData);
 
 	if (res && typeof res === 'object') {
