+++ src/utils/setup-component-properties-for-auto-update.ts
@@ -1,15 +1,16 @@
 import {turnCamelToKebabCasing} from "./turn-camel-to-kebab-casing";
 import {directives} from "../directives";
 import {proxify} from "./proxify";
+import {CWCO} from "../cwco";
 
-export function setupComponentPropertiesForAutoUpdate(component: WebComponent, onUpdate: onUpdateCallback): string[] {
+export function setupComponentPropertiesForAutoUpdate(component: CWCO.WebComponent, onUpdate: CWCO.onUpdateCallback): string[] {
 	const properties: string[] = [];
 
 	for (let property of Object.getOwnPropertyNames(component)) {
 		const attr = turnCamelToKebabCasing(property);
 
 		// ignore private properties and $ properties as well as attribute properties
-		if (!directives.has(property) && !/\$|_/.test(property[0]) && !(component.constructor as WebComponentConstructor).observedAttributes.includes(attr)) {
+		if (!directives.has(property) && !/\$|_/.test(property[0]) && !(component.constructor as CWCO.WebComponentConstructor).observedAttributes.includes(attr)) {
 			// @ts-ignore
 			let value = component[property];
